refactor(daemon): extract runner::registration submodule (#2642)
Move register, register_builtin, register_maintenance_tasks,
register_cron_tasks, and register_knowledge_maintenance_tasks
to runner/registration.rs.
